Routes to Controllers
Broke down routes in jsLand.js to the controllers folder as well as created the controllers bundle for managing controller exports. Add new controller calls to the main js file. All we pass is the server object to set up the routes.
