+++ controllers/controllers.js
@@ -1,0 +1,4 @@
+import { Home } from "./routes/home.js";
+import { Content } from "./routes/content.js";
+
+export { Content, Home };
