--- lobot/css/javascript.css
@@ -1,1 +1,0 @@
-.about-section {}
