+++ lobot/fragments.js
@@ -1,10 +1,10 @@
-import { Head } from "./html/head.js";
-import { Footer } from "./html/footer.js";
-import { ContentGrid } from "./html/contentGrid.js";
-import { Navigation } from "./html/navigation.js";
-import { Home } from "./html/home.js";
+import { Head } from "./resources/html/head.js";
+import { Footer } from "./resources/html/footer.js";
+import { ContentGrid } from "./resources/html/contentGrid.js";
+import { Navigation } from "./resources/html/navigation.js";
+import { Home } from "./resources/html/home.js";
 
 /** Content Articles  **/
-import { TheEventLoop } from "./html/theEventLoop.js";
+import { TheEventLoop } from "./resources/html/theEventLoop.js";
 
 export { ContentGrid, Footer, Head, Home, Navigation, TheEventLoop };
