+++ lobot/resources/js/bootstrap.js
