+++ lobot/resources/js/theme.js
